% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort_layout_brcensus2010}
\alias{mort_layout_brcensus2010}
\title{Mortality data layout}
\format{
data.frame with 7 columns:
  varcode (variable codes),
  vardescr (variable description),
  start_pos (start position of variable information on data files),
  end_pos (final position of variable information on data files),
  interval (interval of variable information on data files),
  decimal (number of decimal positions of variable),
  type (type of variable)
}
\source{
Censo Demográfico 2010. Instituto Brasileiro de Geografia e Estatística (IBGE)
}
\usage{
mort_layout_brcensus2010
}
\description{
Mortality data layout of 2010 Brazilian census
}
\keyword{datasets}
