% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibge_brcensus2010.R
\name{read_ibge_census2010}
\alias{read_ibge_census2010}
\title{Read Brazilian 2010 Census data from IBGE website}
\source{
Censo Demográfico. 2010. Instituto Brasileiro de Geografia e Estatística (IBGE).
}
\usage{
read_ibge_census2010(
  data = "person",
  vars_list = NULL,
  vars_newnames = NULL,
  municode = NULL,
  state = "all",
  save_RData = F,
  save_csv = F,
  format_microdata = F,
  dest_folder = NULL
)
}
\arguments{
\item{data}{data type to be retrieved from IBGE ( person, mortality, household, emigration )}

\item{vars_list}{list of variables (characters) to be retrieved from microdata}

\item{vars_newnames}{(optional) new names for selected variables}

\item{municode}{numerical code of municipality (6 digit)}

\item{state}{state abbreviate name (2 letters character)}

\item{save_RData}{set TRUE to save data in dest_folder as RData}

\item{save_csv}{set TRUE to save data in dest_folder as csv (be careful with sample size)}

\item{format_microdata}{(optional) set TRUE to format microdata according to census layout}

\item{dest_folder}{destination folder to save microdata}
}
\value{
selected states Brazil 2010 census microdata
}
\description{
Fetch Brazilian 2010 Census data from ftp webpage of IBGE
}
\examples{

}
